+++ src/main/java/com/app/group15/UserManagement/ForgetPassword/ForgetPasswordService.java
@@ -20,30 +20,33 @@ public class ForgetPasswordService implements IForgetPasswordService, IForgetPas
 
     private ForgetPasswordAbstractDao forgetPasswordDao;
     private UserDao userDao;
-
+    private String invalidInput = "Invalid invalid";
 
     @Override
     public void generateToken(String bannerId, HttpServletRequest request) {
 
         try {
-            User user = userDao.getUserByBannerId(bannerId);
+            if (ServiceUtility.isNotNull(bannerId) && ServiceUtility.isNotNull(request)) {
+                User user = userDao.getUserByBannerId(bannerId);
 
-            String token = ForgetPasswordUtility.generateForgotPasswordToken();
-            if (forgetPasswordDao.checkIfTokenAlreadyExists(user.getId())) {
-                forgetPasswordDao.deleteForgotPasswordToken(user.getId());
+                String token = ForgetPasswordUtility.generateForgotPasswordToken();
+                if (forgetPasswordDao.checkIfTokenAlreadyExists(user.getId())) {
+                    forgetPasswordDao.deleteForgotPasswordToken(user.getId());
+                }
+
+                forgetPasswordDao.insertForgotPasswordToken(user.getId(), token);
+
+                String urlGenerated = request.getRequestURL().toString();
+                urlGenerated = urlGenerated.substring(0, urlGenerated.lastIndexOf('/'));
+                token = URLEncoder.encode(token, StandardCharsets.UTF_8.toString());
+                urlGenerated += "/auth/validateToken/?to=" + token;
+
+                String subject = "CATME Password Change request";
+                String body = "Click this link to change password: " + urlGenerated;
+                AppConfig.getInstance().getEmailNotifier().sendMessage(user.getEmail(), subject, body);
+            } else {
+                GroupFormationToolLogger.log(Level.SEVERE, invalidInput);
             }
-
-            forgetPasswordDao.insertForgotPasswordToken(user.getId(), token);
-
-            String urlGenerated = request.getRequestURL().toString();
-            urlGenerated = urlGenerated.substring(0, urlGenerated.lastIndexOf('/'));
-            token = URLEncoder.encode(token, StandardCharsets.UTF_8.toString());
-            urlGenerated += "/auth/validateToken/?to=" + token;
-
-            String subject = "CATME Password Change request";
-            String body = "Click this link to change password: " + urlGenerated;
-            AppConfig.getInstance().getEmailNotifier().sendMessage(user.getEmail(), subject, body);
-
         } catch (UnsupportedEncodingException e) {
             GroupFormationToolLogger.log(Level.SEVERE, e.getMessage(), e);
         }
@@ -51,66 +54,96 @@ public class ForgetPasswordService implements IForgetPasswordService, IForgetPas
 
     @Override
     public boolean verifyToken(String token) {
+
         if (ServiceUtility.isNotNull(token)) {
+            if (ServiceUtility.isNotNull(token)) {
 
-            boolean validated = false;
-            Map<String, String> user = forgetPasswordDao.getUserFromToken(token);
-            String tokenGenerationDateTime = user.get("dateTime");
-            Date tokenGenerationDate = null;
-            try {
-                tokenGenerationDate = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss").parse(tokenGenerationDateTime);
-                long minutes = ForgetPasswordUtility.getTimeDifference(tokenGenerationDate);
-                if (minutes > 60) {
-                    forgetPasswordDao.deleteForgotPasswordToken(Integer.parseInt(user.get("id")));
-                    validated = false;
+                boolean validated = false;
+                Map<String, String> user = forgetPasswordDao.getUserFromToken(token);
+                String tokenGenerationDateTime = user.get("dateTime");
+                Date tokenGenerationDate = null;
+                try {
+                    tokenGenerationDate = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss").parse(tokenGenerationDateTime);
+                    long minutes = ForgetPasswordUtility.getTimeDifference(tokenGenerationDate);
+                    if (minutes > 60) {
+                        forgetPasswordDao.deleteForgotPasswordToken(Integer.parseInt(user.get("id")));
+                        validated = false;
+                    }
+                    validated = true;
+                } catch (ParseException e) {
+                    GroupFormationToolLogger.log(Level.SEVERE, e.getMessage(), e);
                 }
-                validated = true;
-            } catch (ParseException e) {
-                GroupFormationToolLogger.log(Level.SEVERE, e.getMessage(), e);
+
+                return validated;
+
+            } else {
+                GroupFormationToolLogger.log(Level.SEVERE, "Invalid input");
+                return false;
             }
-
-            return validated;
-
         } else {
-            GroupFormationToolLogger.log(Level.SEVERE, "Invalid input");
+            GroupFormationToolLogger.log(Level.SEVERE, invalidInput);
             return false;
         }
     }
 
     @Override
     public boolean deleteForgotPasswordToken(int id) {
-        boolean deleted = forgetPasswordDao.deleteForgotPasswordToken(id);
-        return deleted;
+        if (ServiceUtility.isValidInt(id)) {
+            boolean deleted = forgetPasswordDao.deleteForgotPasswordToken(id);
+            return deleted;
+        } else {
+            GroupFormationToolLogger.log(Level.SEVERE, invalidInput);
+            return false;
+        }
     }
 
     @Override
     public Map<String, String> getUserFromToken(String token) {
+        if (ServiceUtility.isNotNull(token)) {
+            Map<String, String> user = forgetPasswordDao.getUserFromToken(token);
+            return user;
+        } else {
+            GroupFormationToolLogger.log(Level.SEVERE, invalidInput);
+            return null;
 
-        Map<String, String> user = forgetPasswordDao.getUserFromToken(token);
-        return user;
+
+        }
     }
 
     @Override
     public boolean updateUserPassword(int id, String newPassword) {
-        boolean updated = forgetPasswordDao.updateUserPassword(id, newPassword);
-        return updated;
+        if (ServiceUtility.isNotNull(newPassword)) {
+            boolean updated = forgetPasswordDao.updateUserPassword(id, newPassword);
+            return updated;
+        } else {
+            GroupFormationToolLogger.log(Level.SEVERE, invalidInput);
+            return false;
+        }
     }
 
     @Override
     public void injectUserDao(UserDao userDao) {
         if (ServiceUtility.isNotNull(userDao)) {
-            this.userDao = userDao;
+            if (ServiceUtility.isNotNull(userDao)) {
+                this.userDao = userDao;
+            } else {
+                GroupFormationToolLogger.log(Level.SEVERE, "userDao injection problems in ForgetPasswordService");
+            }
         } else {
-            GroupFormationToolLogger.log(Level.SEVERE, "userDao injection problems in ForgetPasswordService");
+            GroupFormationToolLogger.log(Level.SEVERE, invalidInput);
         }
     }
 
     @Override
     public void injectForgetPasswordDao(ForgetPasswordAbstractDao forgetPasswordDao) {
         if (ServiceUtility.isNotNull(forgetPasswordDao)) {
-            this.forgetPasswordDao = forgetPasswordDao;
+            if (ServiceUtility.isNotNull(forgetPasswordDao)) {
+                this.forgetPasswordDao = forgetPasswordDao;
+            } else {
+                GroupFormationToolLogger.log(Level.SEVERE, "forgetPasswordDao injection problems in ForgetPasswordService");
+            }
         } else {
-            GroupFormationToolLogger.log(Level.SEVERE, "forgetPasswordDao injection problems in ForgetPasswordService");
+            GroupFormationToolLogger.log(Level.SEVERE, invalidInput);
         }
     }
 }
